split home.index from core and add get-text base file
