+++ gui/slick/js/config/index.js
@@ -1,0 +1,13 @@
+import {getMeta} from '../utils';
+
+const defaults = {
+    srRoot: getMeta('srRoot'),
+    srDefaultPage: getMeta('srDefaultPage'),
+    themeSpinner: getMeta('themeSpinner'),
+    anonURL: getMeta('anonURL'),
+    srPID: getMeta('srPID')
+};
+
+const config = {...defaults};
+
+export default config;
