+++ gui/slick/js/home/index.js
@@ -1,0 +1,5 @@
+import root from './root';
+
+export default {
+    root
+};
