+++ gui/slick/js/utils/get-text.js
@@ -1,0 +1,5 @@
+// @TODO: Move get-text to here
+
+export default {
+    load: () => {}
+};
