+++ gui/slick/js/utils/index.js
@@ -1,10 +1,12 @@
 import getMeta from './get-meta';
+import getText from './get-text';
 import isMeta from './is-meta';
 import metaToBool from './meta-to-bool';
 import notifyModal from './notify-modal';
 
 export {
     getMeta,
+    getText,
     isMeta,
     metaToBool,
     notifyModal
